+++ test/bigtest/interactors/instance-view-page.js
@@ -45,6 +45,7 @@ import MultiColumnListInteractor from '@folio/stripes-components/lib/MultiColumn
   hasHeaderDropdown = isPresent('[class^=DropdownMenuTether---]');
   headerDropdownMenu = new HeaderDropdownMenu();
   itemsList = new MultiColumnListInteractor('#list-items');
+  instanceNotesList = new MultiColumnListInteractor('#list-instance-notes');
   items = collection('#list-items div[class^=mclRow]', Item);
   hasViewHoldingsButton = isPresent('[data-test-view-holdings]');
   clickViewHoldings = clickable('[data-test-view-holdings]');