+++ test/bigtest/tests/instance-view-page-test.js
@@ -114,6 +114,12 @@ describe('InstanceViewPage', () => {
       });
     });
 
+    describe('instance notes list', () => {
+      it('has correct amount of items', () => {
+        expect(InstanceViewPage.instanceNotesList.rowCount).to.be.equal(2);
+      });
+    });
+
     describe('items per holdings', () => {
       describe('sorting', () => {
         describe('compares by `Item: barcode` field', () => {