+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 * Move holdings-record accordion to top of instance display. Refs UIIN-802.
 * Fix display of statistical codes. Fixes UIIN-796.
 * Link to requests by item-id since barcode may not always be present. Fixes UIIN-821.
+* Update the display of list of Instance notes. Refs UIIN-643.
 
 ## [1.12.1](https://github.com/folio-org/ui-inventory/tree/v1.12.1) (2019-09-26)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v1.12.0...v1.12.1)